+++ CHANGELOG.rst
@@ -24,6 +24,8 @@ Core
   (See `#65 <https://gitlab.com/esa/pyxel/issues/65>`_).
 * Added a model for the amplifier crosstalk.
   (See `#116 <https://gitlab.com/esa/pyxel/issues/116>`_).
+* Added ability to load custom QE maps.
+  (See `#117 <https://gitlab.com/esa/pyxel/issues/117>`_).
 
 Others
 ~~~~~~
@@ -38,6 +40,8 @@ Others
   (See `!148 <https://gitlab.com/esa/pyxel/-/merge_requests/148>`_).
 * Fix typo in documentation.
   (See `!149 <https://gitlab.com/esa/pyxel/-/merge_requests/149>`_).
+* Updated documentation according to v0.8.
+  (See `!153 <https://gitlab.com/esa/pyxel/-/merge_requests/153>`_).
 
 
 version 0.7 / 2020-10-22